api/requirements.txt
Flask==0.10.1
virtualenv==1.9.1
six==1.4.1
Flask-Restful==0.2.5
pika==0.9.8
uWSGI==1.9.18.2
PyYAML==3.10
